requirements.txt
openai==1.57.4
typer==0.15.1
jinja2==3.1.4
json5==0.10.0
pyyaml==6.0.2
markdown==3.7
python-docx==1.1.2
beautifulsoup4==4.12.3
gradio==5.9.1
retry==0.9.2